Adhere to abstract adapter initializer
In 8551e64e2411811f26d210601abdba6e13d8798c, rails changed the
[initializer](https://github.com/rails/rails/commit/8551e64e2411811f26d210601abdba6e13d8798c) of the abstract adapter to include a fourth option, while also signaling many of these were deprecated.

This is later used in 009c7e74117690f0dbe200188a929b345c9306c1 by only
passing one argument when creating a [new adapter](https://github.com/rails/rails/commit/009c7e74117690f0dbe200188a929b345c9306c1#diff-0436ae91e81081c450242d55bbcd280e709de611abce3c356a1de462c257ac44R28) in `DatabaseConfig`
and leaning on the default arguments for the other values.

This modifies the signature of the ODBC adapter's initializer to match
and passes the arguments up to super.

This also changes `ODBCAdapter#configure_time_options` so that it only
sets `use_time` if what's passed to it will respond to that method. It
appears that Rails expects to be able to receive a config hash here, and
not always a connection. If we get a hash, it will not respond to
`use_time`.