Test in CI
This includes a Github action that will run the test suite against the
following:

1. Postgres 16
2. Ruby 2.6
3. ActiveRecord 5.2

This introduces the appraisal gem to handle different versions we want
to support. We will expand this to more updated versions of ruby and
rails as we progress in our upkeep of this gem.

Testing against MySql is also an end goal. However, I'm starting with
only testing Postgres as the intent here is to confirm we are still
adhering to the ActiveRecord API as we proceed through newer and newer
versions. As such, testing against one database engine should be
sufficient to meet that goal.