Version 6.0.0
Given that this is dropping support for a version of ActiveRecord, the
version of the adapter is changing to the next major version.

Especially given the lack of a release artifact, I think this is
particularly important to stay on top of.

Note that the version numbering here is not meant to mirror the version
of ActiveRecord. That is a coincidence for this version. A version that
drops support for AR 6.0 will not be gem version 6.1. It will go to 7.x.