Limited Support for ActiveRecord 7.0
We know we have complicated compatibility issues with AR 7.2, and are
trying to work our way up to that point by getting the test suite we're
using to assist with troubleshooting that. We currently use this for
connectivity with Snowflake, and are using it with AR 7.1 in production.
Our use case is pretty limited to sending raw queries and handling the
casting on our own for results.

This transitions us to testing against ActiveRecord 7.0 to move us
closer to making those changes to support 7.2. It does so by making the
MINIMAL changes for our use case.

⚠️ This INTENTIONALLY limits the usefulness of this gem for generic ODBC
connectivity going forward. ⚠️

To support AR 7.0, this:

1. Uses the `type_cast` method for handling binds, rather than the
   `_type_cast` method. The underscore method was inlined in [c5bf2b4736f2ddafbc477af5b9478dd7143e5466](https://github.com/rails/rails/commit/c5bf2b4736f2ddafbc477af5b9478dd7143e5466).
   This requires changes to `PostgreSQLODBCAdapter#type_cast` to
   adhere to the expected public API, in terms of arguments accepted.
   Because `type_cast` no longer accepts a column, as of [2d821ef0c5356d860eba2d2341b4db71b67f6594](https://github.com/rails/rails/commit/2d821ef0c5356d860eba2d2341b4db71b67f6594),
   this custom method isn't useful, as it would return `super` without a column.
2. Remove timezone deprecation warning. When running tests against 7.0,
   we see the following deprecation:

   ```

   DEPRECATION WARNING: ActiveRecord::Base.default_timezone is deprecated and will be removed in Rails 7.1.
Use `ActiveRecord.default_timezone` instead.
   ```

A limitation that this currently has is that columns retrieved from the
database are not appropriately being cast to their proper data types. To
get around this in the test suite, I've modified the model definition in
the test helper to explicitly use the Attributes API to tell the AR User
class that letters is an integer.

You can see this in `CalculationsTest#test_average`, which would fail
without that change. It would be attempting to call `round` on a string,
which fails. This intentionally avoids resolving that problem so we can
proceed on updating for our very limited use case on more recent
versions of ActiveRecord.

There are known mapping changes that occurred between 6.1 and 7.0.
However, none of those changes have been beneficial for resolving this
bug.

In [2ec207520389f4a6acb393be45bc73cc3815ad76](https://github.com/rails/rails/commit/2ec207520389f4a6acb393be45bc73cc3815ad76),
the `type_map` is moved to be private API in the connection adapter.

In [d79fb963603658117fd1d639976c375ea2a8ada3](https://github.com/rails/rails/commit/d79fb963603658117fd1d639976c375ea2a8ada3),
the type map is defined statically for improved memory performance.

We can also see how another fork modified this code to support [AR 7.1](https://github.com/springbuk/odbc_adapter/commit/60a6c9dc8a5707af560506a83f2715b7ed9a59aa)
(which this is not supporting yet).