ActiveRecord 7.1 compatibility
This increases the minimum version of ActiveRecord to be used against
this to be 7.1. This is no longer compatible with 7.0 after these
changes.

The following large-scale modifications have been made to support 7.1:

1. Rename `@connection` instance variable to `@raw_connection`.

In [d86fd6415c0dfce6fadb77e74696cf728e5eb76b](https://github.com/rails/rails/commit/d86fd6415c0dfce6fadb77e74696cf728e5eb76b),
the instance variable `@connection` is renamed to `@raw_connection`.
This makes the same change throughout the adapter.

> This is unfortunately going to force a change into all external adapter
> subclasses, but I'm going to be making other changes too, so it's not
> the worst of them... and it's just so confusing at the moment to have
> "@connection" sometimes mean the AR adapter and sometimes the inner
> connection, depending on exactly which class we're in.

2. Modify adapter constructor arguments

In [8551e64e2411811f26d210601abdba6e13d8798c](https://github.com/rails/rails/commit/8551e64e2411811f26d210601abdba6e13d8798c),
the `AbstractAdapter`'s initializer is changed to, going forward, expect
only a configuration hash. However, it still has support for the prior
connection attributes. They have been soft-deprecated.

The initializer of this adapter has been changed to adhere to that set
of arguments while also adding the database metadata, as it previously
did.

Note that this adapter still expects the "old" style of providing all
the arguments, and not only the config hash. This means that when that
transition eventually happens and the initializer hard-warns,
deprecates, and eventually removes the old style, we will be forced to
address that then.

3. Define `reconnect` rather than `reconnect!`

In [02f5de1bc93407e878cf4d92e2caf46d4c6e5409](https://github.com/rails/rails/commit/02f5de1bc93407e878cf4d92e2caf46d4c6e5409),
the `AbstractAdapter#reconnect!` method expects for a `reconnect` method
to exist on any implementations of the adapter. This renames what was
the `reconnect!` method in our adapter to instead be the `reconnect`
method that `reconnect!` calls.

4. Define `DatabaseStatements#internal_exec_query`

In [942785fb8b03772dde338cc241af8e5de4d2efbc](https://github.com/rails/rails/commit/942785fb8b03772dde338cc241af8e5de4d2efbc)
ActiveRecord keeps the `DatabaseStatements#exec_query` public API.
However, it delegates the implementation to `internal_exec_query`, as do
other methods. The `internal_exec_query` method must be defined by each
adapter. This change does so.

5. Modify `DatabaseStatements#prepared_binds`

This is the most curious of these changes, in that I cannot provide
justifiable source code explanation on the change. In ActiveRecord 7.0,
when binding parameters, the parameter to bind would be an instance of
`ActiveRecord::Relation::QueryAttribute`. In 7.1, it appears to be the
raw bound value. So, attempting to call `value_for_database` on a symbol
or a string or other database does not succeed, as it doesn't respond to
that method.

This is at least the case for binding params into queries communicating
with Postgres, as this CI test suite currently does.

To avoid this, rather than definitely calling `value_for_database`, we
check to see if each bind responds to the method, and if so, then we
retrieve `value_for_database`. Otherwise, we pass the bind value
directly to be type cast.

I spent a good amount of time looking at the Arel visitors to understand
the AST differences between 7.0 and 7.1. I could clearly see there
were a different set of visitors used for the same query for this
binding. However, it was not clear to be WHY that was the case, or what
changes caused that. As such, this feels like not getting at the root
cause; however, it does allow us to proceed.

This has been tested internally against an application connecting with
Snowflake in addition to passing the tests on CI.