Support nil database_metadata in adapter constructor
The addition of the database_metadata attribute is not part of the
standard constructor that ActiveRecord may expect. It's added by this
adapter specifically so `ActiveRecord::Base.odbc_connection` can pass
the already constructed database metadata to the constructor.

Because the rest of ActiveRecord will not send it, this defaults the
attribute to nil. When it's not provided, we construct it based on the
connection.