Extract connection setup
ActiveRecord 7.2 includes changes with how an adapter is created. As a
result, it's likely that the work that's done in
`ActiveRecord::Base.odbc_connection` will move or be re-used in
additional locations to support 7.2.

To get ahead of this, the logic to build the connection and modify the
configuration is extracted to a separate class that is individually
tested.