Drop Ruby 3.2 from the testing matrix
We want to test against newer versions of Ruby that are in use. Ruby 3.2
is still in [security maintenance](https://www.ruby-lang.org/en/downloads/branches/)
and not scheduled to reach EOL until 2026. However, this is getting
ahead of that to reduce the amount of coverage in our CI suite because
we as an organization do not have a defined need to test against Ruby
3.2 anymore.